Fixed bug in Glue.hpp that would run into asserts at corners.
The bug would occur for scalar and continuous discretizations.
Therein, if one identifies mutliple DOFs at a corner, both DOFs
of the last pair would already have been assigned a master, and
asserts in lines 319 and 324 would be triggered. I added an if-
statement that checks for this case, and if it comes up verifies
that both DOFs have been assigned the same master.