Fixes to make Bembel compatible to VisualStudio
This includes replacing "and" by "&&", "not" by "!" and
"M_PI" by "BEMBEL_PI", as well as one minor fix in ElementTree.
Therein, a reference to possibly non-allocated memory was created,
and a check for allocation was performed afterwards. Although
this was technically correct, since the case of non-allocation was
treated specifically, this would trigger asserts in Microsofts
implementation of std::vector when compiled with debug flags.
Now, we check for allocation before.